move SiteRegisterClient legacy code under gnr.web.daemon, to be replaced by new packages.
Refactored references accordingly
